lets do what Tgstation is doing; database schema things; fix balloon alert to viewers; add tabletop anchor offset; I think I will do... papers later.
