+++ SQL/database_changelog.md
@@ -5,12 +5,44 @@ Make sure to also update `DB_MAJOR_VERSION` and `DB_MINOR_VERSION`, which can be
 The latest database version is 2.0; The query to update the schema revision table is:
 
 ```sql
-INSERT INTO `schema_revision` (`major`, `minor`) VALUES (2, 0);
+INSERT INTO `schema_revision` (`major`, `minor`) VALUES (2, 1);
 ```
 
 
 In any query remember to add a prefix to the table names if you use one.
 
+-----------------------------------------------------
+Version 2.1 1 October 2025, by Flleeppyy
+Alter `library` table to add , add `library_action` table.
+
+```sql
+ALTER TABLE `library`
+DROP FOREIGN KEY `fk_rails_53d51ce16a`,
+DROP INDEX `index_library_on_author_id`,
+CHANGE COLUMN `author` `author` VARCHAR(45) NOT NULL ,
+CHANGE COLUMN `title` `title` VARCHAR(45) NOT NULL ,
+CHANGE COLUMN `content` `content` TEXT NOT NULL ,
+CHANGE COLUMN `category` `category` ENUM('Any','Fiction','Non-Fiction','Adult','Reference','Religion') NOT NULL ,
+CHANGE COLUMN `author_id` `ckey` VARCHAR(32) NOT NULL DEFAULT 'LEGACY' ,
+CHANGE COLUMN `created_at` `datetime` DATETIME NOT NULL ,
+DROP COLUMN `updated_at`,
+ADD COLUMN `round_id_created` INT(11) UNSIGNED NULL AFTER `deleted`,
+ADD INDEX `deleted_idx` (`deleted` ASC),
+ADD INDEX `idx_lib_id_del` (`id` ASC, `deleted` ASC),
+ADD INDEX `idx_lib_del_title` (`deleted` ASC, `title` ASC),
+ADD INDEX `idx_lib_search` (`deleted` ASC, `author` ASC, `title` ASC, `category` ASC);
+
+CREATE TABLE `library_action` (
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+  `book` int(10) unsigned NOT NULL,
+  `reason` longtext DEFAULT NULL,
+  `ckey` varchar(32) NOT NULL DEFAULT '',
+  `datetime` datetime NOT NULL DEFAULT current_timestamp(),
+  `action` varchar(11) NOT NULL DEFAULT '',
+  `ip_addr` int(10) unsigned NOT NULL,
+  PRIMARY KEY (`id`)
+) ENGINE=InnoDB AUTO_INCREMENT=27 DEFAULT CHARSET=utf8mb4;
+```
 -----------------------------------------------------
 Version 2.0 5 September 2025, by Flleeppyy
 Add `messages` table
@@ -41,7 +73,6 @@ CREATE TABLE `messages` (
   KEY `idx_msg_type_ckey_time_odr` (`type`,`targetckey`,`timestamp`, `deleted`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
 ```
------------------------------------------------------
 -----------------------------------------------------
 Version 1.0 16 March 2025, by Flleeppyy
 Add `byond_build` and `byond_version` to the `connection_log` table.
