+++ cev_eris.dme
@@ -1888,7 +1888,6 @@
 #include "code\modules\library\book_info.dm"
 #include "code\modules\library\bookcase.dm"
 #include "code\modules\library\lib_machines.dm"
-#include "code\modules\library\lib_readme.dm"
 #include "code\modules\lighting\lighting_area.dm"
 #include "code\modules\lighting\lighting_atom.dm"
 #include "code\modules\lighting\lighting_corner.dm"
