+++ code/modules/library/admin_only.dm
@@ -2,6 +2,8 @@
 #define BOOK_ADMIN_RESTORE "undeleted"
 #define BOOK_ADMIN_REPORT "reported"
 
+#warn figure out why the fuck does this not work AT ALL
+
 /obj/machinery/computer/libraryconsole/admin_only_do_not_map_in_you_fucker
 	interface_type = "LibraryAdmin"
 	/// When a user clicks view, do we display the raw text, or process it with markdown
@@ -80,7 +82,7 @@
 	page_count = round(max(bookcount - 1, 0) / BOOKS_PER_PAGE) //This is just floor()
 	search_page = clamp(search_page, 0, page_count)
 
-/obj/machinery/computer/libraryconsole/admin_only_do_not_map_in_you_fucker/ui_status(mob/user)
+/obj/machinery/computer/libraryconsole/admin_only_do_not_map_in_you_fucker/ui_status(mob/user, datum/ui_state/state)
 	if(!check_rights_for(user.client, R_BAN))
 		return UI_CLOSE
 	if(!SSdbcore.Connect())
@@ -140,7 +142,7 @@
 	.["view_raw"] = view_raw
 	.["show_deleted"] = show_deleted
 	var/list/histories = list()
-	for(var/id as anything in book_history)
+	for(var/id in book_history)
 		var/list/insert = list()
 		for(var/datum/book_history_entry/entry in book_history[id])
 			insert += list(entry.serialize())
@@ -339,7 +341,7 @@
 		ui.set_autoupdate(FALSE) // Nothing is changing here brother
 		ui.open()
 
-/datum/admin_book_viewer/ui_status(mob/user)
+/datum/admin_book_viewer/ui_status(mob/user, datum/ui_state/state)
 	if(!check_rights_for(user.client, R_BAN))
 		return UI_CLOSE
 	return UI_INTERACTIVE
