+++ code/modules/library/lib_machines.dm
@@ -25,6 +25,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 	icon_keyboard = null
 	circuit = /obj/item/electronics/circuitboard/libraryconsole
 	desc = "Checked out books MUST be returned on time."
+	anchored_tabletop_offset = 8
 	///The current book id we're searching for
 	var/book_id = null
 	///The current title we're searching for
@@ -79,7 +80,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 	data["params_changed"] = params_changed
 	return data
 
-/obj/machinery/computer/libraryconsole/ui_act(action, params)
+/obj/machinery/computer/libraryconsole/ui_act(action, list/params, datum/tgui/ui, datum/ui_state/state)
 	. = ..()
 	if(.)
 		return
@@ -390,6 +391,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 				data["cache_content"] = scan.cache.get_content()
 
 		if(LIBRARY_PRINT)
+			data["posters"] = list()
 			for(var/poster_name in SSlibrary.printable_posters)
 				data["posters"] += poster_name
 
@@ -411,7 +413,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 		scanner = WEAKREF(foundya)
 		return foundya
 
-/obj/machinery/computer/libraryconsole/bookmanagement/ui_act(action, params)
+/obj/machinery/computer/libraryconsole/bookmanagement/ui_act(action, list/params, datum/tgui/ui, datum/ui_state/state)
 	//The parent call takes care of stuff like searching, don't forget about that yeah?
 	. = ..()
 	if(.)
@@ -428,10 +430,12 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 			var/id = params["book_id"]
 			inventory -= id
 			inventory_update()
+			update_static_data_for_all_viewers()
 			return TRUE
 		if("switch_inventory_page")
 			inventory_page = sanitize_page_input(params["page"], inventory_page, inventory_page_count)
 			inventory_update()
+			update_static_data_for_all_viewers()
 			return TRUE
 		if("checkout")
 			var/list/available = list()
@@ -488,24 +492,20 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 		// if("news_post")
 		// 	// We grey out the button UI-side, but let's just be safe to guard against spammy spammers.
 		// 	if(!COOLDOWN_FINISHED(src, newscaster_cooldown))
-		// 		say_quote("Not enough time has passed since the last news post. Please wait.")
+		// 		say("Not enough time has passed since the last news post. Please wait.")
 		// 		return
 		// 	if(!GLOB.news_network)
-		// 		say_quote("No news network found on station. Aborting.")
-		// 	var/channelexists = FALSE
-		// 	for(var/datum/feed_channel/feed in GLOB.news_network.network_channels)
-		// 		if(feed.channel_name == LIBRARY_NEWSFEED)
-		// 			channelexists = TRUE
-		// 			break
-		// 	if(!channelexists)
+		// 		say("No news network found on station. Aborting.")
+		// 	var/datum/feed_channel/library_channel = GLOB.news_network.network_channels_by_name[LIBRARY_NEWSFEED]
+		// 	if(isnull(library_channel))
 		// 		GLOB.news_network.create_feed_channel(LIBRARY_NEWSFEED, "Library", "The official station book club!", null)
 
 		// 	var/obj/machinery/libraryscanner/scan = get_scanner()
 		// 	if(!scan)
-		// 		say_quote("No nearby scanner detected. Aborting.")
+		// 		say("No nearby scanner detected. Aborting.")
 		// 		return
 		// 	GLOB.news_network.submit_article(scan.cache.content, "[scan.cache.author]: [scan.cache.title]", LIBRARY_NEWSFEED, null)
-		// 	say_quote("Upload complete. Your uploaded title is now available on station newscasters.")
+		// 	say("Upload complete. Your uploaded title is now available on station newscasters.")
 		// 	COOLDOWN_START(src, newscaster_cooldown, NEWSCASTER_COOLDOWN)
 		// 	return TRUE
 		if("print_book")
@@ -674,6 +674,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 	icon = 'icons/obj/library.dmi'
 	icon_state = "bigscanner"
 	desc = "It's an industrial strength book scanner. Perfect!"
+	// circuit = /obj/item/circuitboard/machine/libraryscanner
 	density = TRUE
 	var/obj/item/book/held_book
 	///Our scanned-in book
@@ -746,6 +747,7 @@ GLOBAL_VAR_INIT(library_table_modified, 0)
 	icon = 'icons/obj/library.dmi'
 	icon_state = "binder"
 	desc = "Only intended for binding paper products."
+	// circuit = /obj/item/circuitboard/machine/bookbinder
 	density = TRUE
 
 	/// Are we currently binding a book?
