+++ tgui/packages/tgui/package.json
@@ -13,7 +13,7 @@
     "marked": "^4.3.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "tgui-core": "^2.1.1",
+    "tgui-core": "^3.3.1",
     "tgui-dev-server": "workspace:*"
   },
   "devDependencies": {
